content/svg/content/test/test_dataTypes.html

<!DOCTYPE html>
<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=437448
-->
<head>
  <title>Test for Bug 437448</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=437448">Mozilla Bug 437448</a>
<p id="display"></p>
<div id="content" style="display: none"></div>

<iframe id="svg" src="dataTypes-helper.svg"></iframe>

<pre id="test">
<script class="testbody" type="application/javascript">
SimpleTest.waitForExplicitFinish();

function runTests()
{
  var doc = $("svg").contentWindow.document;
  var filter = doc.getElementById("filter");
  var convolve = doc.getElementById("convolve");
  var blur = doc.getElementById("blur");
  var marker = doc.getElementById("marker");

  // class attribute
  filter.setAttribute("class", "foo");
  is(filter.getAttribute("class"), "foo", "class attribute");
  is(filter.className.baseVal, "foo", "className baseVal");
  is(filter.className.animVal, "foo", "className animVal");
  filter.className.baseVal = "bar";
  is(filter.getAttribute("class"), "bar", "class attribute");
  is(filter.className.baseVal, "bar", "className baseVal");
  is(filter.className.animVal, "bar", "className animVal");
  filter.removeAttribute("class");
  is(filter.hasAttribute("class"), false, "class attribute");
  is(filter.className.baseVal, "", "className baseVal");
  is(filter.className.animVal, "", "className animVal");

  // length attribute

  marker.setAttribute("markerWidth", "12.5");
  is(marker.markerWidth.baseVal.value, 12.5, "length baseVal");
  is(marker.markerWidth.animVal.value, 12.5, "length animVal");

  var baseMarkerWidth = marker.markerWidth.baseVal;
  var animMarkerWidth = marker.markerWidth.animVal;
  marker.setAttribute("markerWidth", "15.5");
  is(baseMarkerWidth.value, 15.5, "length baseVal");
  is(animMarkerWidth.value, 15.5, "length animVal");

  marker.markerWidth.baseVal.value = 7.5;
  is(marker.markerWidth.animVal.value, 7.5, "length animVal");
  is(marker.getAttribute("markerWidth"), "7.5", "length attribute");

  // number attribute

  convolve.setAttribute("divisor", "12.5");
  is(convolve.divisor.baseVal, 12.5, "number baseVal");
  is(convolve.divisor.animVal, 12.5, "number animVal");

  convolve.divisor.baseVal = 7.5;
  is(convolve.divisor.animVal, 7.5, "number animVal");
  is(convolve.getAttribute("divisor"), "7.5", "number attribute");

  // number-optional-number attribute

  blur.setAttribute("stdDeviation", "20.5");
  is(blur.stdDeviationX.baseVal, 20.5, "number-optional-number first baseVal");
  is(blur.stdDeviationX.animVal, 20.5, "number-optional-number first animVal");
  is(blur.stdDeviationY.baseVal, 20.5, "number-optional-number second baseVal");
  is(blur.stdDeviationY.animVal, 20.5, "number-optional-number second animVal");

  blur.stdDeviationX.baseVal = 8.5;
  is(blur.stdDeviationX.animVal, 8.5, "number-optional-number first animVal");
  is(blur.stdDeviationY.animVal, 20.5, "number-optional-number second animVal");
  is(blur.getAttribute("stdDeviation"), "8.5, 20.5", "number-optional-number attribute");

  blur.stdDeviationY.baseVal = 8.5;
  is(blur.getAttribute("stdDeviation"), "8.5", "number-optional-number attribute");

  blur.setStdDeviation(24.5, 0.5);
  is(blur.stdDeviationX.baseVal, 24.5, "integer-optional-integer first baseVal");
  is(blur.stdDeviationX.animVal, 24.5, "integer-optional-integer first animVal");
  is(blur.stdDeviationY.baseVal, 0.5, "integer-optional-integer second baseVal");
  is(blur.stdDeviationY.animVal, 0.5, "integer-optional-integer second animVal");

  // integer attribute

  convolve.setAttribute("targetX", "12");
  is(convolve.targetX.baseVal, 12, "integer baseVal");
  is(convolve.targetX.animVal, 12, "integer animVal");
  convolve.targetX.baseVal = 7;
  is(convolve.targetX.animVal, 7, "integer animVal");
  is(convolve.getAttribute("targetX"), "7", "integer attribute");

  // integer-optional-integer attribute

  filter.setAttribute("filterRes", "100");
  is(filter.filterResX.baseVal, 100, "integer-optional-integer first baseVal");
  is(filter.filterResX.animVal, 100, "integer-optional-integer first animVal");
  is(filter.filterResY.baseVal, 100, "integer-optional-integer second baseVal");
  is(filter.filterResY.animVal, 100, "integer-optional-integer second animVal");

  filter.filterResX.baseVal = 50;
  is(filter.filterResX.animVal, 50, "integer-optional-integer first animVal");
  is(filter.filterResY.animVal, 100, "integer-optional-integer second animVal");
  is(filter.getAttribute("filterRes"), "50, 100", "integer-optional-integer attribute");

  filter.filterResY.baseVal = 50;
  is(filter.getAttribute("filterRes"), "50", "integer-optional-integer attribute");

  filter.setFilterRes(80, 90);
  is(filter.filterResX.baseVal, 80, "integer-optional-integer first baseVal");
  is(filter.filterResX.animVal, 80, "integer-optional-integer first animVal");
  is(filter.filterResY.baseVal, 90, "integer-optional-integer second baseVal");
  is(filter.filterResY.animVal, 90, "integer-optional-integer second animVal");

  // angle attribute

  marker.setAttribute("orient", "90deg");
  is(marker.orientAngle.baseVal.value, 90, "angle baseVal");
  is(marker.orientAngle.animVal.value, 90, "angle animVal");

  var baseAngle = marker.orientAngle.baseVal;
  var animAngle = marker.orientAngle.animVal;
  marker.setAttribute("orient", "45deg");
  is(baseAngle.value, 45, "angle baseVal");
  is(animAngle.value, 45, "angle animVal");

  marker.orientAngle.baseVal.value = 30;
  is(marker.orientAngle.animVal.value, 30, "angle animVal");
  is(marker.getAttribute("orient"), "30deg", "angle attribute");

  // boolean attribute

  convolve.setAttribute("preserveAlpha", "false");
  is(convolve.preserveAlpha.baseVal, false, "boolean baseVal");
  is(convolve.preserveAlpha.animVal, false, "boolean animVal");
  convolve.preserveAlpha.baseVal = true;
  is(convolve.preserveAlpha.animVal, true, "boolean animVal");
  is(convolve.getAttribute("preserveAlpha"), "true", "boolean attribute");

  // enum attribute

  convolve.setAttribute("edgeMode", "wrap");
  is(convolve.edgeMode.baseVal, 2, "enum baseVal");
  is(convolve.edgeMode.animVal, 2, "enum animVal");
  convolve.edgeMode.baseVal = 1;
  is(convolve.edgeMode.animVal, 1, "enum animVal");
  is(convolve.getAttribute("edgeMode"), "duplicate", "enum attribute");

  // string attribute

  convolve.setAttribute("result", "foo");
  is(convolve.result.baseVal, "foo", "string baseVal");
  is(convolve.result.animVal, "foo", "string animVal");
  convolve.result.baseVal = "bar";
  is(convolve.result.animVal, "bar", "string animVal");
  is(convolve.getAttribute("result"), "bar", "string attribute");

  // preserveAspectRatio attribute

  marker.setAttribute("preserveAspectRatio", "xMinYMid slice");
  is(marker.preserveAspectRatio.baseVal.align, 5, "preserveAspectRatio.align baseVal");
  is(marker.preserveAspectRatio.animVal.align, 5, "preserveAspectRatio.align animVal");
  is(marker.preserveAspectRatio.baseVal.meetOrSlice, 2, "preserveAspectRatio.meetOrSlice baseVal");
  is(marker.preserveAspectRatio.animVal.meetOrSlice, 2, "preserveAspectRatio.meetOrSlice animVal");
  marker.preserveAspectRatio.baseVal.align = 3;
  is(marker.preserveAspectRatio.animVal.align, 3, "preserveAspectRatio animVal");
  is(marker.preserveAspectRatio.animVal.meetOrSlice, 2, "preserveAspectRatio.meetOrSlice animVal");
  marker.preserveAspectRatio.baseVal.meetOrSlice = 1;
  is(marker.preserveAspectRatio.animVal.align, 3, "preserveAspectRatio animVal");
  is(marker.preserveAspectRatio.animVal.meetOrSlice, 1, "preserveAspectRatio.meetOrSlice animVal");
  is(marker.getAttribute("preserveAspectRatio"), "xMidYMin meet", "preserveAspectRatio attribute");

  var basePreserveAspectRatio = marker.preserveAspectRatio.baseVal;
  var animPreserveAspectRatio = marker.preserveAspectRatio.animVal;
  marker.setAttribute("preserveAspectRatio", "xMaxYMid slice");
  is(basePreserveAspectRatio.align, 7, "preserveAspectRatio.align baseVal");
  is(animPreserveAspectRatio.align, 7, "preserveAspectRatio.align animVal");
  is(basePreserveAspectRatio.meetOrSlice, 2, "preserveAspectRatio.meetOrSlice baseVal");
  is(animPreserveAspectRatio.meetOrSlice, 2, "preserveAspectRatio.meetOrSlice animVal");

  // viewBox attribute
  var baseViewBox = marker.viewBox.baseVal;
  var animViewBox = marker.viewBox.animVal;
  is(baseViewBox.x, 0, "viewBox baseVal");
  is(animViewBox.x, 0, "viewBox baseVal");
  is(baseViewBox.y, 0, "viewBox baseVal");
  is(animViewBox.y, 0, "viewBox baseVal");
  is(baseViewBox.width, 0, "viewBox baseVal");
  is(animViewBox.width, 0, "viewBox baseVal");
  is(baseViewBox.height, 0, "viewBox baseVal");
  is(animViewBox.height, 0, "viewBox baseVal");
  baseViewBox.x = 10;
  baseViewBox.y = 11;
  baseViewBox.width = 12;
  baseViewBox.height = 13;
  is(marker.getAttribute("viewBox"), "10 11 12 13", "viewBox attribute");

  marker.setAttribute("viewBox", "1 2 3 4");
  is(marker.viewBox.baseVal.x, 1, "viewBox.x baseVal");
  is(marker.viewBox.animVal.x, 1, "viewBox.x animVal");
  is(marker.viewBox.baseVal.y, 2, "viewbox.y baseVal");
  is(marker.viewBox.animVal.y, 2, "viewbox.y animVal");
  is(marker.viewBox.baseVal.width, 3, "viewbox.width baseVal");
  is(marker.viewBox.animVal.width, 3, "viewbox.width animVal");
  is(marker.viewBox.baseVal.height, 4, "viewbox.heigth baseVal");
  is(marker.viewBox.animVal.height, 4, "viewbox.heigth animVal");
  marker.viewBox.baseVal.x = 5;
  is(marker.viewBox.animVal.x, 5, "viewBox.x animVal");
  marker.viewBox.baseVal.y = 6;
  is(marker.viewBox.animVal.y, 6, "viewBox.y animVal");
  marker.viewBox.baseVal.width = 7;
  is(marker.viewBox.animVal.width, 7, "viewBox.width animVal");
  marker.viewBox.baseVal.height = 8;
  is(marker.viewBox.animVal.height, 8, "viewBox.height animVal");
  is(marker.getAttribute("viewBox"), "5 6 7 8", "viewBox attribute");
  marker.removeAttribute("viewBox");
  is(marker.hasAttribute("viewBox"), false, "viewBox hasAttribute");

  SimpleTest.finish();
}

window.addEventListener("load", runTests, false);
</script>
</pre>
</body>
</html>
